add repositories

+++ backend/src/main/kotlin/com/bftcom/backend/entity/BookWork.kt
@@ -2,6 +2,6 @@ package com.bftcom.backend.entity
 
 data class BookWork(
 	override var id: Long = 0,
-	val bookId: Int,
-	val workId: Int
+	val bookId: Long,
+	val workId: Long
 ) : Entity
